packages/hardware_node/requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.1; python_version >= '3.6'
aiosignal==1.2.0; python_version >= '3.6'
appdirs==1.4.4
asttokens==2.0.8
async-timeout==4.0.2; python_version >= '3.6'
attrs==22.1.0; python_version >= '3.5'
backcall==0.2.0
charset-normalizer==2.1.1; python_version >= '3.6'
cobs==1.2.0
crc8==0.1.0
decorator==5.1.1; python_version >= '3.5'
executing==1.0.0
frozenlist==1.3.1; python_version >= '3.7'
idna==3.3; python_version >= '3.5'
ipython==8.5.0; python_version >= '3.8'
jedi==0.18.1; python_version >= '3.6'
matplotlib-inline==0.1.6; python_version >= '3.5'
msgpack==1.0.4
multidict==6.0.2; python_version >= '3.7'
# numpy==1.23.3
odrive==0.6.1.post0
parso==0.8.3; python_version >= '3.6'
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
prompt-toolkit==3.0.31; python_full_version >= '3.6.2'
ptyprocess==0.7.0
pure-eval==0.2.2
pyelftools==0.29
pygments==2.13.0; python_version >= '3.6'
pyserial==3.5
pyusb==1.2.1; python_version >= '3.6'
# pyyaml==6.0
setuptools==65.3.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
stack-data==0.5.0
traitlets==5.3.0; python_version >= '3.7'
wcwidth==0.2.5
yarl==1.8.1; python_version >= '3.7'
